working conversion to typescript
